Incorporated parse arguments; Looped the heat diffusion over n_its; Changed some variables name; Added comments

+++ A4/heat_calc.cl
@@ -2,7 +2,7 @@ __kernel
 void
 heat_diff(
 	__global const float *a,
-	const double c,
+	const float c,
 	__global float *n,
 	int width_a
 	)
